csci5410-summer-23-sdp34-main/Notification_module/TeamInvite/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub==2.8.0
boto3
google-cloud-firestore
